+++ R/bmerge.R
@@ -1,41 +1,22 @@
 
-bmerge <- function(i, x, leftcols, rightcols, roll, rollends, nomatch, mult, ops, verbose)
+bmerge <- function(i, x, icols, xcols, roll, rollends, nomatch, mult, ops, verbose)
 {
-  # TO DO: rename leftcols to icols, rightcols to xcols
-  # TO DO: xo could be moved inside Cbmerge
-  # bmerge moved to be separate function now that list() doesn't copy in R
-  # types of i join columns are promoted to match x's types (with warning or verbose)
+  callersi = i
+  i = shallow(i)  # important: i is already passed in as a shallow copy from [.data.table because [.data.table
+                       # builds the answer from (potentially) coerced i, other than factor columns for which the
+                       # original levels in i are retained in the final answer.
+  x = shallow(x)       # if any coercions are done for x, we must be careful not to change the caller's view of x.
+  # careful to only plonk syntax (full column) on i/x from now on (otherwise user's i and x would change);
+  #   this is why shallow() is very importantly internal only, currently.
 
-  # Important that i is already passed in as a shallow copy, due to these coercions for factors.
-  # i.e. bmerge(i<-shallow(i),...)
-  # The caller ([.data.table) then uses the coerced columns to build the output
+  #  DELETE ... resetifactor = NULL     # Keep track of any factor to factor/character join cols (only time we keep orig)
 
-  # careful to only plonk syntax (full column) on i from now on (otherwise i would change)
-  # TO DO: enforce via .internal.shallow attribute and expose shallow() to users
-  # This is why shallow() is very importantly internal only, currently.
-  x     = shallow(x)
-  origi = shallow(i)      # Needed for factor to factor/character joins, to recover the original levels
-                          # Otherwise, types of i join columns are anyways promoted to match x's
-                          # types (with warning or verbose)
-  resetifactor = NULL     # Keep track of any factor to factor/character join cols (only time we keep orig)
-
-  ## function to determine type of columns.
-  ## Important not to use typeof() since integer and factor have same type (integer)
-  ## important to test for is.double instead of is.numeric because of bug with IDate (test 346)
-  getClass <- function(x){
-    if(is.logical(x))                 out <- "logical"
-    else if(is.integer(x))            out <- "integer"
-    else if(base::is.double(x)){ ## base::is.double(integer64) is TRUE, while bit64::is.double(integer64) is FALSE!
-      if(inherits(x, "integer64"))    out <- "integer64" ## is.numeric returns TRUE on integer64 columns
-      ## distinguish integer values with storage mode double from real numeric
-      ## values that can't be coerced to integer without loss of precision
-      else if(isReallyReal(x))        out <- "reallyDouble"
-      else                            out <- "intAsDouble"
-    }
-    else if(is.character(x))          out <- "character"
-    else if(is.factor(x))             out <- "factor"
-    else                              out <- "other"
-    return(out)
+  getClass <- function(x) {
+    ans = typeof(x)
+    if      (ans=="integer") { if (is.factor(x))             ans = "factor"    }
+    else if (ans=="double")  { if (inherits(x, "integer64")) ans = "integer64" }
+    # do not call isReallyReal(x) yet because if both types are double, we don't need to coerce even if one or both sides are int-as-double
+    ans
   }
 
   ## function or coercing a vector to a new type. Depending on the source and target type,
@@ -82,119 +63,116 @@ bmerge <- function(i, x, leftcols, rightcols, roll, rollends, nomatch, mult, ops
   ## rows mark the column type in i, columns the column type in x
   ## possible types are: logical, integer, intAsDouble, reallyDouble, character, factor, integer64
 
-  ct = as.matrix(read.table(text="
-               logical integer intAsDouble reallyDouble character factor integer64 other
-  logical            y      ci          ci           ci         e      e        ci     e
-  integer            e       y          ci           ci         e      e        ci     e
-  intAsDouble        e      ci           y            y         e      e        ci     e
-  reallyDouble       e      cx           y            y         e      e        cx     e
-  character          e       e           e            e         y     ci         e     e
-  factor             e       e           e            e        ci      y         e     e
-  integer64          e      ci          ci           ci         e      e         y     e
-  other              e       e           e            e         e      e         e     e
-  "))
-
-  for (a in seq_along(leftcols)) {
+  supported = c("logical", "integer", "double", "character", "factor", "integer64")
+  for (a in seq_along(icols)) {
     # This loop does the following:
     # - check that join columns have compatible types
     # - do type coercions if necessary
     # - special support for joining factor columns
     # Note that if i is keyed, if this coerces, i's key gets dropped and the key may not be retained
-    lc = leftcols[a]   # i   # TO DO: rename left and right to i and x
-    rc = rightcols[a]  # x
-    icnam = names(i)[lc]
-    xcnam = names(x)[rc]
-    myXclass = getClass(x[[rc]])
-    myIclass = getClass(i[[lc]])
-    myXtype  = myXclass
-    if(myXtype %chin% c("intAsDouble", "reallyDouble")) myXtype = "double"
-    if(myXtype == "other") myXtype = paste0(class(x[[rc]]), collapse = ",")
-    myItype  = myIclass
-    if(myItype %chin% c("intAsDouble", "reallyDouble")) myItype = "double"
-    if(myItype == "other") myItype = paste0(class(i[[lc]]), collapse = ",")
-    coercionStrategy <- ct[myIclass, myXclass]
-
-    if(coercionStrategy == "y"){
-      ## nothing to be done, but no 'next' since the factor related stuff below needs to be executed
-    }
-    else if(coercionStrategy == "e"){
-      stop(sprintf("Incompatible types: %s (%s) and %s (%s)",
-                   paste0("x.", xcnam), myXtype, paste0("i.", icnam), myItype))
-    }
-    else if(coercionStrategy %chin% c("ci", "ciw")){
-      ## coerce i[[lc]] to same class as x[[rc]]
-      if (verbose) {cat(sprintf("Coercing %s column %s to %s to match type of %s.",
-                                myItype, paste0("i.'", icnam, "'"), myXtype, paste0("x.'", xcnam, "'"))); flush.console()}
-      if(coercionStrategy == "ciw"){
-        warning(sprintf("Coercing %s column %s to %s to match type of %s.",
-                        myItype, paste0("i.'", icnam, "'"), myXtype, paste0("x.'", xcnam, "'")))
-      }
-      if(myXtype == "factor"){
-        ## special treatment due to factor levels, see #499 and #945
-        newval = coerceClass(origi[[lc]], to = myXtype) ## will do mode() coercion if possible to retain attributes
-        set(origi, j=lc, value=newval)
-        # TO DO: we need a way to avoid copying 'value' for internal purposes
-        # that would allow setting: set(i, j=lc, value=origi[[lc]]) without resulting in a copy.
-        # until then using 'val <- origi[[lc]]' below to avoid another copy.
+    ic = icols[a]
+    xc = xcols[a]
+    xclass = getClass(x[[xc]])
+    iclass = getClass(i[[ic]])
+    if (!xclass %chin% supported) stop("x.", names(x)[xc]," is type ", xclass, " which is not supported by data.table join")
+    if (!iclass %chin% supported) stop("i.", names(i)[ic]," is type ", iclass, " which is not supported by data.table join")
+    if (xclass=="factor" || iclass=="factor") {
+      if (roll!=0.0 && a==length(icols))
+        stop("Attempting roll join on factor column when joining x.",names(x)[xc]," to i.",names(i)[ic],". Only integer, double or character colums may be roll joined.")
+      if (xclass=="factor" && iclass=="factor") {
+        # Levels are matched to levels here, then bmerge.c continues as-if integer
+        # Retain original levels of i's factor column (important when NAs, see tests 687 and 688).  Related fixes: #499 and #945
+        # delete ... resetifactor = c(resetifactor,lc)
+        #            val = origi[[lc]] # note: using 'origi' here because set(..., value = .) always copies '.', we need a way to avoid it in internal cases.
+        #            lx = levels(x[[xc]])
+        #            li = levels(val)
+        if (verbose) cat("Matching i.",names(i)[ic]," factor levels to x.",names(x)[xc]," factor levels.\n",sep="")
+        set(i, j=ic, value=chmatch(levels(i[[ic]]), levels(x[[xc]]), nomatch=0L)[i[[ic]]])
+        next
+        #            levels(newfactor) = lx
+        #            class(newfactor) = "factor"
+        #            set(i, j=lc, value=newfactor)
       } else {
-        newval = coerceClass(i[[lc]], to = myXtype) ## will do mode() coercion if possible to retain attributes
-        set(i, j=lc, value=newval)
+        if (xclass=="character") {
+          if (verbose) cat("Coercing factor column i.",names(i)[ic]," to character to match type of x.",names(x)[xc],".\n",sep="")
+          val = as.character(i[[ic]])
+          set(i, j=ic, value=val)
+          set(callersi, j=ic, value=val)
+          next
+        } else if (iclass=="character") {
+          if (verbose) cat("Matching character column i.",names(i)[ic]," to factor levels in x.",names(x)[xc],".\n",sep="")
+          set(i, j=ic, value=chmatch(i[[ic]], levels(x[[xc]]), nomatch=0L))
+          next
+        }
       }
-      myItype = myXtype
     }
-    else if(coercionStrategy %chin% c("cx", "cxw")){
-      ## coerce x[[rc]] to same class as i[[lc]]
-      if (verbose) {cat(sprintf("Coercing %s column %s to %s to match type of %s.",
-                                myXtype, paste0("x.'", xcnam, "'"), myItype, paste0("i.'", icnam, "'"))); flush.console()}
-      if(coercionStrategy == "cxw"){
-        warning(sprintf("Coercing %s column %s to %s to match type of %s.",
-                        myXtype, paste0("x.'", xcnam, "'"), myItype, paste0("i.'", icnam, "'")))
+    if (xclass == iclass) {
+      if (verbose) cat("i.",names(i)[ic],"has same type (",xclass,") as x.",names(x)[xc],". No coercion needed.")
+      next
+    }
+    if (xclass=="character" || iclass=="character" ||
+        xclass=="logical" || iclass=="logical" ||
+        xclass=="factor" || iclass=="factor") {
+      stop("Incompatible join types: x.", names(x)[xc], " (",xclass,") and i.", names(i)[ic], " (",iclass,")")
+    }
+    if (xclass=="integer64" || iclass=="integer64") {
+      if (xclass=="integer64") { w=i; wc=ic; wclass=iclass; } else { w=x; wc=xc; wclass=xclass; }  # w is which to coerce
+      if (wclass=="integer" || (wclass=="double" && !isReallyReal(w[[wc]]))) {
+        val = as.integer64(w[[wc]])
+        set(w, j=wc, value=val)
+        if (wclass==iclass) set(callersi, j=wc, value=val)
+      } else stop("cannot coerce reallyreal to integer64")
+    } else {
+      # just integer and double left
+      if (iclass=="double") {
+        if (!isReallyReal(i[[ic]])) {
+          # common case of ad hoc user-typed integers missing L postfix joining to correct integer keys
+          # we've always coerced to int and returned int, for convenience.
+          if (verbose) cat("Coercing double column i.",names(i)[ic]," which contains integers to type integer to match type of x.",names(x)[xc],".\n",sep="")
+          val = as.integer(i[[ic]])
+          set(callersi, j=ic, value=val)       # change the shallow copy of i up in [.data.table to reflect in the result
+          set(i, j=ic, value=val)  # change local shallow copy too to apply in the Cmerge call below
+        }
+        else stop("Cannot coerce double to int")
+      } else {
+        val = as.double(i[[ic]])
+        set(i, j=ic, value=val)
+        set(callersi, j=ic, value=val)
       }
-      newval = coerceClass(x[[rc]], to = myItype) ## will do mode() coercion if possible to retain attributes
-      set(x, j=rc, value=newval)
-      myXtype = myItype
-    }
-    else stop("Internal error in bmerge: unknown type coercion strategy.")
-
-    ## now take care about factor columns.
-    if(myXtype == "factor"){
-      if(myItype != "factor") stop("Internal error in bmerge: at this point, myItype should be factor.")
-      # levels of factors have to be treated properly when coercing
-      # Retain original levels of i's factor columns in factor to factor joins (important when NAs,
-      # see tests 687 and 688).
-      # Moved it outside of 'else' to fix #499 and #945.
-      resetifactor = c(resetifactor,lc)
-      if (roll!=0.0 && a==length(leftcols)) stop("Attempting roll join on factor column x.",xcnam,". Only integer, double or character colums may be roll joined.")   # because the chmatch on next line returns <strike>NA</strike> <new>0</new> for missing chars in x (rather than some integer greater than existing). Note roll!=0.0 is ok in this 0 special floating point case e.g. as.double(FALSE)==0.0 is ok, and "nearest"!=0.0 is also true.
-      val = origi[[lc]] # note: using 'origi' here because set(..., value = .) always copies '.', we need a way to avoid it in internal cases.
-      lx = levels(x[[rc]])
-      li = levels(val)
-      newfactor = chmatch(li, lx, nomatch=0L)[val] # fix for #945, a hacky solution for now.
-      levels(newfactor) = lx
-      class(newfactor) = "factor"
-      set(i, j=lc, value=newfactor)
     }
   }
+  # browser()
 
-  ## after all modifications of i, check if i has a proper key on all leftcols
-  io <- identical(leftcols, head(chmatch(key(i), names(i)), length(leftcols)))
+#      stop(sprintf("Incompatible types: %s (%s) and %s (%s)",
+#                   paste0("x.", xcnam), myXtype, paste0("i.", icnam), myItype))
 
-  ## after all modifications of x, check if x has a proper key on all rightcols.
+#      if (verbose) {cat(sprintf("Coercing %s column %s to %s to match type of %s.",
+#                                myItype, paste0("i.'", icnam, "'"), myXtype, paste0("x.'", xcnam, "'"))); flush.console()}
+
+#        warning(sprintf("Coercing %s column %s to %s to match type of %s.",
+#                        myItype, paste0("i.'", icnam, "'"), myXtype, paste0("x.'", xcnam, "'")))
+
+
+  ## after all modifications of i, check if i has a proper key on all icols
+  io <- identical(icols, head(chmatch(key(i), names(i)), length(icols)))
+
+  ## after all modifications of x, check if x has a proper key on all xcols.
   ## If not, calculate the order. Also for non-equi joins, the order must be calculated.
   non_equi = which.first(ops != 1L) # 1 is "==" operator
   if (is.na(non_equi)) {
     # equi join. use existing key (#1825) or existing secondary index (#1439)
-    if (identical(rightcols, head(chmatch(key(x), names(x)), length(rightcols)))) {
+    if (identical(xcols, head(chmatch(key(x), names(x)), length(xcols)))) {
       xo = integer(0L)
       if (verbose) cat("on= matches existing key, using key\n")
     } else {
       xo <- NULL
       if (isTRUE(getOption("datatable.use.index"))) {
-        xo = getindex(x, names(x)[rightcols])
+        xo = getindex(x, names(x)[xcols])
         if (verbose && !is.null(xo)) cat("on= matches existing index, using index\n")
       }
       if (is.null(xo)) {
         if (verbose) {last.started.at=proc.time(); flush.console()}
-        xo = forderv(x, by = rightcols)
+        xo = forderv(x, by = xcols)
         if (verbose) {cat("Calculated ad hoc index in",timetaken(last.started.at),"\n"); flush.console()}
         # TODO: use setindex() instead, so it's cached for future reuse
       }
@@ -210,10 +188,10 @@ bmerge <- function(i, x, leftcols, rightcols, roll, rollends, nomatch, mult, ops
     if (roll != FALSE) stop("roll is not implemented for non-equi joins yet.")
     if (verbose) {last.started.at=proc.time();cat("  forder took ... ");flush.console()}
     # TODO: could check/reuse secondary indices, but we need 'starts' attribute as well!
-    xo = forderv(x, rightcols, retGrp=TRUE)
+    xo = forderv(x, xcols, retGrp=TRUE)
     if (verbose) {cat(timetaken(last.started.at),"\n"); flush.console()}
     xg = attr(xo, 'starts')
-    resetcols = head(rightcols, non_equi-1L)
+    resetcols = head(xcols, non_equi-1L)
     if (length(resetcols)) {
       # TODO: can we get around having to reorder twice here?
       # or at least reuse previous order?
@@ -223,31 +201,33 @@ bmerge <- function(i, x, leftcols, rightcols, roll, rollends, nomatch, mult, ops
       if (verbose) {cat("done in",timetaken(last.started.at),"\n"); flush.console()}
     } else resetlen = integer(0L)
     if (verbose) {last.started.at=proc.time();cat("  Generating non-equi group ids ... ");flush.console()}
-    nqgrp = .Call(Cnestedid, x, rightcols[non_equi:length(rightcols)], xo, xg, resetlen, mult)
+    nqgrp = .Call(Cnestedid, x, xcols[non_equi:length(xcols)], xo, xg, resetlen, mult)
     if (verbose) {cat("done in",timetaken(last.started.at),"\n"); flush.console()}
     if (length(nqgrp)) nqmaxgrp = max(nqgrp) # fix for #1986, when 'x' is 0-row table max(.) returns -Inf.
     if (nqmaxgrp > 1L) { # got some non-equi join work to do
       if ("_nqgrp_" %in% names(x)) stop("Column name '_nqgrp_' is reserved for non-equi joins.")
       if (verbose) {last.started.at=proc.time();cat("  Recomputing forder with non-equi ids ... ");flush.console()}
       set(nqx<-shallow(x), j="_nqgrp_", value=nqgrp)
-      xo = forderv(nqx, c(ncol(nqx), rightcols))
+      xo = forderv(nqx, c(ncol(nqx), xcols))
       if (verbose) {cat("done in",timetaken(last.started.at),"\n"); flush.console()}
     } else nqgrp = integer(0L)
     if (verbose) cat("  Found", nqmaxgrp, "non-equi group(s) ...\n")
   }
 
   if (verbose) {last.started.at=proc.time();cat("Starting bmerge ...");flush.console()}
-  ans = .Call(Cbmerge, i, x, as.integer(leftcols), as.integer(rightcols), io, xo, roll, rollends, nomatch, mult, ops, nqgrp, nqmaxgrp)
+  ans = .Call(Cbmerge, i, x, as.integer(icols), as.integer(xcols), io, xo, roll, rollends, nomatch, mult, ops, nqgrp, nqmaxgrp)
   if (verbose) {cat("done in",timetaken(last.started.at),"\n"); flush.console()}
+  # TO DO: xo could be moved inside Cbmerge
 
-  # in the caller's shallow copy,  see comment at the top of this function for usage
+  # delete ... in the caller's shallow copy,  see comment at the top of this function for usage
   # We want to leave the coercions to i in place otherwise, since the caller depends on that to build the result
-  if (length(resetifactor)) {
-    for (ii in resetifactor)
-      set(i,j=ii,value=origi[[ii]])
-    if (haskey(origi))
-      setattr(i, 'sorted', key(origi))
-  }
+  #if (length(resetifactor)) {
+  #  for (ii in resetifactor)
+  #    set(i,j=ii,value=origi[[ii]])
+  #  if (haskey(origi))
+  #    setattr(i, 'sorted', key(origi))
+  #}
+
   ## add xo for further use
   ans$xo <- xo
   return(ans)